+++ common/src/main/java/dev/terminalmc/signedit/mixin/input/ContainerEventHandlerMixin.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2026 TerminalMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.terminalmc.signedit.mixin.input;
+
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import dev.terminalmc.signedit.SignEdit;
+import net.minecraft.client.gui.components.events.ContainerEventHandler;
+import net.minecraft.client.input.KeyEvent;
+import org.spongepowered.asm.mixin.Mixin;
+
+import static dev.terminalmc.signedit.config.Config.options;
+
+@Mixin(ContainerEventHandler.class)
+@SuppressWarnings("JavadocReference")
+public interface ContainerEventHandlerMixin {
+
+    /**
+     * Removes released keys from the 'down' list.
+     *
+     * @see LocalPlayerMixin#onOpenTextEdit
+     * @see dev.terminalmc.signedit.mixin.edit.AbstractSignEditScreenMixin#wrapKeyPressed
+     */
+    @WrapMethod(method = "keyReleased")
+    default boolean wrapKeyReleased(KeyEvent event, Operation<Boolean> original) {
+        if (options().blockMovementKeys) {
+            SignEdit.downKeys.removeIf(keyMapping -> keyMapping.matches(event));
+        }
+        return original.call(event);
+    }
+}